refactor(docs): select dx
