+++ apps/docs/content/components/select/disabled.ts
@@ -1,43 +1,7 @@
-const data = `export const animals = [
-  {key: "cat", label: "Cat"},
-  {key: "dog", label: "Dog"},
-  {key: "elephant", label: "Elephant"},
-  {key: "lion", label: "Lion"},
-  {key: "tiger", label: "Tiger"},
-  {key: "giraffe", label: "Giraffe"},
-  {key: "dolphin", label: "Dolphin"},
-  {key: "penguin", label: "Penguin"},
-  {key: "zebra", label: "Zebra"},
-  {key: "shark", label: "Shark"},
-  {key: "whale", label: "Whale"},
-  {key: "otter", label: "Otter"},
-  {key: "crocodile", label: "Crocodile"}
-];`;
-
-const App = `import {Select, SelectItem} from "@nextui-org/react";
-import {animals} from "./data";
-
-export default function App() {
-  return (
-    <Select
-      isDisabled
-      label="Favorite Animal"
-      placeholder="Select an animal"
-      defaultSelectedKeys={["cat"]}
-      className="max-w-xs"
-    >
-      {animals.map((animal) => (
-        <SelectItem key={animal.key}>
-          {animal.label}
-        </SelectItem>
-      ))}
-    </Select>
-  );
-}`;
+import App from "./disabled.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
-  "/data.js": data,
 };
 
 export default {
