+++ apps/docs/content/components/select/multiple-controlled.ts
@@ -1,81 +1,12 @@
-const data = `export const animals = [
-  {key: "cat", label: "Cat"},
-  {key: "dog", label: "Dog"},
-  {key: "elephant", label: "Elephant"},
-  {key: "lion", label: "Lion"},
-  {key: "tiger", label: "Tiger"},
-  {key: "giraffe", label: "Giraffe"},
-  {key: "dolphin", label: "Dolphin"},
-  {key: "penguin", label: "Penguin"},
-  {key: "zebra", label: "Zebra"},
-  {key: "shark", label: "Shark"},
-  {key: "whale", label: "Whale"},
-  {key: "otter", label: "Otter"},
-  {key: "crocodile", label: "Crocodile"}
-];`;
-
-const App = `import {Select, SelectItem} from "@nextui-org/react";
-import {animals} from "./data";
-
-export default function App() {
-  const [values, setValues] = React.useState(new Set(["cat", "dog"]));
-
-  return (
-    <div className="flex w-full max-w-xs flex-col gap-2">
-      <Select
-        label="Favorite Animal"
-        selectionMode="multiple"
-        placeholder="Select an animal"
-        selectedKeys={values}
-        className="max-w-xs"
-        onSelectionChange={setValues}
-      >
-        {animals.map((animal) => (
-          <SelectItem key={animal.key}>
-            {animal.label}
-          </SelectItem>
-        ))}
-      </Select>
-      <p className="text-small text-default-500">Selected: {Array.from(values).join(", ")}</p>
-    </div>
-  );
-}`;
-
-const AppTs = `import {Select, SelectItem, Selection} from "@nextui-org/react";
-import {animals} from "./data";
-
-export default function App() {
-  const [values, setValues] = React.useState<Selection>(new Set(["cat", "dog"]));
-
-  return (
-    <div className="flex w-full max-w-xs flex-col gap-2">
-      <Select
-        label="Favorite Animal"
-        selectionMode="multiple"
-        placeholder="Select an animal"
-        selectedKeys={values}
-        className="max-w-xs"
-        onSelectionChange={setValues}
-      >
-        {animals.map((animal) => (
-          <SelectItem key={animal.key}>
-            {animal.label}
-          </SelectItem>
-        ))}
-      </Select>
-      <p className="text-small text-default-500">Selected: {Array.from(values).join(", ")}</p>
-    </div>
-  );
-}`;
+import App from "./multiple-controlled.raw.jsx?raw";
+import AppTs from "./multiple-controlled.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
-  "/data.js": data,
 };
 
 const reactTs = {
   "/App.tsx": AppTs,
-  "/data.ts": data,
 };
 
 export default {
